refactor(ai-engine): split logic_auditor_agent into logic_auditor subpackage (#1708)
- Extract ai-engine/agents/logic_auditor_agent.py (32K) into ai-engine/agents/logic_auditor/ subpackage
- Modules: pattern_detector, behavior_validator, semantic_diff, complexity_analyzer, audit_reporter
- Reduce original file to re-export stub
- Verified audit_reporter scoring tests still pass after split (48 logic_auditor tests, 50 with audit keyword, all pass)

Fixes #1708
